made a seperate folder for notebook management, seperated the running script from the reference file, wrote a quick version of just importing the config file, importing relevant packages, and getting a return from the html parser, which ran. modified the .toml file to include under optional dependencies one for myself, wasn't sure if the right structure. also, first commit.
